pyproject.toml
[tool.poetry]
name = "helper-methods-stocks"
version = "0.1.0"
description = ""
authors = ["Andrew_mac <andrewkravchuk97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^2.1.3"
requests = "~2.32.0"
pyspark = "*"
websockets = "^14.0"
ipykernel = "^6.29.5"
azure-identity = "^1.19.0"
azure-keyvault-secrets = "^4.9.0"
delta-spark = "^3.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.requires-plugins]
poetry-plugin-export = ">=1.8"